BUG#22256752 - PURGE THREAD DIES: FAILING ASSERTION: COL_LEN - 8 > 1 || LEN < 256
To store blob data, we would check the max blob data length according to
blob type, for example, TINYBLOB requries 1 byte to store the length,
and 255 bytes for the data itself, and BLOB requires 2 bytes for length, etc.

Under this constraint, to get the computed BLOB for virtual column,
we always pass a buffer with fixed length according to row format,
this is OK for BLOB, but not OK for TINYBLOB. Thus it results in the
failing assertion.

The fix is that we should create and pass a buffer of proper length for
all kinds of BLOBs, to support TINYBLOB, we just need a buffer of 255 bytes.

RB: 11197
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>