Bug#21452916  SILENT FAILURE TO START IF MYSQL.GTIDS_EXECUTED GETS HA_ERR_LOCK_WAIT_TIMEOUT
If a server is crashed after preparing a XA transaction and right before
committing the XA transaction, which is used to modify the system
gtid_executed table explicitly by users, then the following recovery
is aborting due to an innodb_lock_wait_timeout error when it is
reading the system gtid_executed table.

To fix the problem, we do not allow users to modify the
mysql.gtid_executed table explicitly by a XA transaction.