WL#8706 Persistent storage of Histogram data
This worklog covers the work needed to store histogram data persistent. This is
needed so that histogram data does not have to be created each time the server
starts.

We have created one new system table for storing histogram statistics;
mysql.column_stats. The histograms will be stored in a JSON column, which
gives us a very flexible and powerful way of storing histograms.