Bug#19031409 REVISIT WHAT HAPPENS IF ALLOCATION OF THE JOIN BUFFER FAILS
Doc: "an out-of-memory failure in join buffer allocation could lead
to wrong results for multi-table queries".

Some code was trying to gracefully handle OOM in join buffer allocation,
and it was tested in i_main.join_cache. But the result in that test
was actually incorrect (one excess row). Fixed the relevant code:
now, instead of trying to still use the already-allocated buffers of
previous tables in the join order (which led to wrong results due to
dependencies between the buffers), it disables them all.
Note that in this unlikely scenario, the query will not necessarily get
the same query plan as when join buffering is disabled before optimizing
the query.  The cost calculations would be different, and index
selection/join order might be different.  Hence, the query plan may be
far from optimal if one is not able to create join buffers.