Bug#22297983 HEX() FUNCTION CONVERTS DECIMAL TO DOUBLE AND THUS RETURNS WRONG RESULT
The patch for "Bug #9854 hex() and out of range handling"
introduced val_real() rather than val_int() in the implementation
of the "hex" function for decimals. Alas, it loses precision for large values.

Solution: revert the patch for Bug #9854.
Let val_int() do the job, it will do correct rounding for
floats/decimals, and will not lose precision for large values.

Change the test case for Bug #9854 to show that "hex" and "conv"
actually do the same thing for "out of range" input data.