Bug#22340858 CRASH AT GET_INTERVAL_VALUEITEMINTERVAL_TYPE
Problem: Parsing the output of ST_GeometryType as a DATETIME value
when the default character set is UTF32 causes the server to crash.

Item_func_geometry_type::val_str_ascii() copies the hard-coded ASCII
encoded name of the type into the result string, but claims that the
character set is default_charset().

The calling function expects an ASCII value and calls my_isdigit() on
the result, which segfaults since treating the 4 byte character set as
ASCII causes it to dereference a null pointer.

Fix: Set Latin 1 as the character set of the return value from
ST_GeometryType.