Bug#22222013: ASSERTION `! IS_SET() || M_CAN_OVERWRITE_STATUS' FAILED IN ::SET_ERROR_STATUS
Problem: When there are multiple parser errors there is an assertion.
This is because in the existing code while copying parser diagnostic
area DA to sql DA (not sure what wording to use for thd->m_stmt_da)
we try to set the error status when there was already an error set
which leads to assertion.
Fix: Fix is to check if Da has error status or not and then only set
the parser error.