Bug#22223202: Query with double nested subquery much slower in 5.7
In 5.7, we unconditionally merge a derived table or view reference
into the outer query when possible and when not told otherwise
by setting optimizer switch derived_merge=off.

However, some queries with derived tables benefit from better
performance when the subqueries are materialized. One such possible
case is a derived table with dependent subqueries in the select list,
used as the inner table of a left outer join. Such tables will always
be read as many times as there are qualifying rows in the outer table,
and the select list subqueries are evaluated for each row combination.
The select list subqueries are evaluated the same number of times
also with join buffering enabled, even though the table then only will
be read once.

Hence, applications may suffer from worse performance due to this.
We will therefore revert some of the extra optimizations added in 5.7:
Derived tables containing dependent subqueries in their SELECT list,
will not be candidate for merging any more. We continue to merge
derived tables with non-dependent subqueries, since such subqueries
are executed only once, regardless of how many row combinations they
are evaluated for.

We should consider to re-enable this feature when we have implemented
hints to directly impact whether to merge derived tables. We should
also consider to improve the cost optimizer to calculate one set of
plans with derived tables materialized and another set of plans
with merging, and then pick the best of the best.