Bug#22280780: MAIN.JSON FAILS ON FREEBSD AND OS X
Part II.

The main.json test fails with debug builds on FreeBSD and OS X because
Json_dom::parse() runs out of stack space when processing a deeply
nested JSON document. This patch reduces the stack usage in the
function, so that the test again passes on those platforms.

It reduces stack space usage of the Json_dom::parse() function in
two ways:

1) By reducing the number of std::unique_ptr objects used in
Json_dom::parse().

2) By reducing the size of the json_binary::Value objects.

Details:

On the platforms where the stack overflow is seen, std::unique_ptr
seems to be the big contributor. The problem has only been seen with
debug builds on FreeBSD and OS X, and only when built with clang. The
problem was not seen before bug#22254630 replaced std::auto_ptr with
std::unique_ptr.

Json_dom::parse() is a recursive function, and it pushes a unique_ptr
onto the stack every time it descends into the next level of the JSON
document. The patch changes this, so that it only has one unique_ptr
which holds the root of the document. When descending down the
document tree, it does not add a new unique_ptr on every level.
Instead, it adds the Json_dom nodes for subdocuments to their parent
Json_object or Json_array immediately when they are created. Since
Json_object and Json_array destroy all their children in their
destructors, and the root of the document is stored in a unique_ptr,
we know that the entire document tree will be freed if
Json_dom::parse() returns abnormally.

Another contributor to the stack usage, is the json_binary::Value
class. There is at least one such object on the stack at each
recursion level, and the size of the objects is greater than it needs
to be.

The patch reduces the size of the json_binary::Value objects from 56
bytes to 24 bytes on the affected platforms. The size reduction is
achieved by the following changes:

- Use uint32 instead of size_t for m_length and m_element_count. Since
  the binary format specifies that binary JSON documents must be
  smaller than 4GB, uint32 is sufficient for storing the size.

- Declare that the underlying type of the
  json_binary::Value::enum_type enumeration is uint8, so that the
  m_type member variable can fit in a single byte.

- Make a union of the member variables m_data, m_int_value and
  m_double_value. A json_binary::Value object will only use one of
  those members, depending on which JSON type it represents, so there
  is no need to set aside space for all of them at once.

- Reorder the member variables so that the smaller ones come at the
  end and can fill "holes" caused by alignment.