Bug#22280780: MAIN.JSON FAILS ON FREEBSD AND OS X
Part I.

The main.json test fails because of a stack overflow when processing a
deep JSON document in debug builds on FreeBSD and OS X.

The recursive function that fails, was recently rewritten to use
std::unique_ptr instead of std::auto_ptr (bug#22254630). This appears
to have increased the footprint on the stack significantly on some
platforms. It has only been seen with debug builds on FreeBSD and OS
X, and only when compiling with clang.

This patch makes the failing recursive function (Json_dom::parse)
check if it is about to run out of stack space by calling
check_stack_overrun(), so that the statement fails gracefully with a
proper error message instead of exiting the server when it is short on
stack.