BUG#21798355 SLOW STARTUP OF 5.7.X SLAVE WITH RELAY_LOG_RECOVERY = ON              AND MANY RELAY LOGS
Problem
-------

The slave initialization was scanning relay log files to update the
Retrieved_Gtid_Set and the transaction parser state even when the
relay_log_recovery option was set.

Analysis
--------

During the slave initialization, it opens the relay log to compute the
Retrieved_Gtid_Set and to set the transaction parser in the correct
state in order to keep track of already replicated GTID sets.

The initialization procedure was performing the above actions even when
the relay_log_recovery option was set.

When the relay_log_recovery option is set, the slave I/O thread should
not use the current relay log files and will base it initial state on
the SQL thread state. If the I/O thread will not rely on relay log
files, it is a waste of time to compute the Retrieved_Gtid_Set and the
transaction parser state when the relay_log_recovery option is set.

Fix
---

The Retrieved_Gtid_Set and the transaction parser states are now
initialized only if the relay_log_recovery option is not set.