Bug #21796658: SERVER HANG DURING UNINSTALL AND INSTALL AUDIT_LOG PLUGIN
Reporting errors to the client can result in logging audit events.
So if this is done while holding the LOCK_plugin mutex the thread is going
to hang since logging audit events may result in trying to acquire
LOCK_plugin.

Fixed by making sure that LOCK_plugin is not held when errors are reported.