Bug#22016556 INNODB LOOKS FOR BUFFER POOL FILE NAME IN '/' IF INNODB_DATA_HOME_DIR IS EMPTY
Issue:
======
If the server is started with the following parameter set in cnf file -
"innodb_data_home_dir =", to specify absolute paths for the data files
listed in the innodb_data_file_path value, then the server looks for the
buffer pool dump file in the root directory and throws an error.

Fix:
====
The directory path of the buffer pool dump file is handled such that the
server creates and looks for the buffer pool dump file at the right place.

RB: 11081
Reviewed-by: Satya Bodapati <satya.bodapati@oracle.com>
Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>