Bug#22376872 FIX ERRORS REPORTED BY UBSAN
Add --ubsan option to mtr, to search for ubsan reports in the server logs.

mysqltest.cc
negation of -2147483648 cannot be represented in type 'int'
Fix: make the entire expression unsigned

sql_auth_cache.h
downcast of address which does not point to an object of type 'GRANT_TABLE'
Fix: remove the cast

item_buff.cc
downcast of address which does not point to an object of type 'Item_field'
Fix: remove the cast

item_timefunc.cc
downcast of address which does not point to an object of type 'Item_date_add_interval'
Fix: move code to where it belongs: after call to Item_func::eq()

item_xmlfunc.cc
downcast of address which does not point to an object of type 'Item_func'
Fix: dont cast to Item_func, use Item, the function may simply be 'true' or 'false'

sql_base.cc
downcast of address which does not point to an object of type 'Item_direct_view_ref'
note: object is of type 'Item_field'
Fix: use Item_ident, which is a common base class for the two classes.

sql_resolver.cc
downcast of address which does not point to an object of type 'Item_field'
note: object is of type 'Item_direct_view_ref'
Fix: call real_item() to get to the actual field.

gis0sea.cc
member call on address which does not point to an object of type 'rw_lock_t'
note: object has invalid vptr
Fix: the target memory was not an object, it was raw memory.
Use copy CTOR rather than operator=()