Bug#21277472  MYSQLD DOES NOT HANDLE NDBAPI TIMEOUTS CORRECTLY
Fix for autotest failure 'testDict -nBug58277'. The previous
patch removed some error-handling in the getIndex operation
which was overwriting all errors to error 4243. However, an
error code overwrite is required for error 723. Added an
exception for error 723 "table not found" so that it can be
mapped to error 4243 "index not found".