WL#11065: Backport JSON functions from 8.0 to 5.7 (part 7)
The following bug fix is backported to 5.7 from trunk:

Bug#22253965: UPDATE CRASHES IN SORTCMP

Problem:

A query that executed a JSON function which raised an error, caused
the server to exit.

Analysis:

If an error happens when Item_cache_json::cache_value() evaluates the
expression that should be cached, it immediately returns false to
indicate that it could not cache the expression. It does not touch the
null_value flag, so it will still tell the nullability of the previous
value that was cached.

Item_cache_json::val_str() returns NULL if the cache couldn't get
populated. It does not touch the null_value flag either.

Because of this, we could get in a situation where
Item_cache_json::val_str() returns NULL because an error happened
during evaluation of the expression to cache, but the null_value flag
indicates that a non-NULL value was returned. This inconsistency could
confuse callers of Item_cache_json::val_str(). For example:

Query_result_max_min_subquery::cmp_str() uses the null_value flag to
determine if val_str() returned a NULL value. If null_value is false,
it believes that it is safe to dereference the pointer returned by
val_str(). Because of the inconsistency, the returned pointer could be
NULL even though null_value is false, and this would cause cmp_str()
to dereference the NULL pointer and the server would exit.

Solution:

Make Item_cache_json::cache_json() update the null_value flag also in
the situations where an error is raised during evaluation of the
expression.

Change-Id: Ie02c538913d565327a76879da628c18aa648912c