Patch for Bug#22204186, and duplicates Bug#17674771 , bug#19537961
Bug#22204186  NDB REPLICATION : READ ONLY TABLES AND NO BINLOGGING
Bug#17674771  NDB TABLES LET READ-ONLY INDEFINITELY AFTER MYSQLD RECONNECT TO THE CLUSTER
bug#19537961  ERROR HANDLING INSIDE BINLOG INJECTOR THREAD

Root cause of this bug is a race condition between the Util thread
regularly(~1s) trying to set up the binlog, and the binlog thread
itself doing failure handling:

The binlog thread is supposed to drop all eventOperations
it subscribe to when it receive a failure event. When all failure
events are handled, it assume to find that getEventOperation()
on both the 's_ndb' and 'i_ndb' connection will return NULL which
will be its condition to restart itself. (Inbetween the first
failure event and successfully restarting, all tables will
be in a raed only mode)

On the other hand, the Util-thread will regularly call
ndb_binlog_setup() which will create the schema distr. tables
if required, and the event subscriptions on these. Iff this happened
when the binlog thread had only partially handled the failure events,
we never reached its restart condition of all 'getEventOperations()==NULL'.
This the binlog is never restarted, no gap event written to the log,
and it stayed permanently in read only mode.

The fix is to ensure that the '*schema_ndb' is set to 'NULL'
when the event dropping the 'ndb_schema' table is handled.
ndb_binlog_setup() already contains a guard preventing it from
completing its task when the schema_ndb object does not exist.
Thus there will be no new eventOps created in the 's_ndb',
and the remaining failures will eventually be handled such
that also the 'i_ndb' event ops are dropped. The binlog thread
will then restart.

Upon restart the 'schema_ndb' handle is recreated and
the eventOp on the 'ndb_schema' table set up. This allows
ndb_binlog_setup() to go forward and complete its work.

This patch also contains some cleanup of:

- Use 'NULL' instead of '0' when assigning to pointers
- Replace 'BCCC_exit' with 'BCCC_starting' as that was the real
  use of it.
- Replace the meaningless usage of 'binlog_thread_state' in
  boolean expressions, where 'binlog_thread_state != BCCC_running'
  should have been used.
- Simplified some binlog-loop conditions where the above
  binlog_thread_state pattern were used.
- Removed a DBUG code block which effectively were dead
  code, as binlog_thread_state is known to be 'BCCC_running'
  at this place.

NOTE to above: In the main binlog loop binlog_thread_state
 can only transition from running -> restart. The previous
 usage of 'exit' seems to have been obsoleted at some point
 (guess) by is_stop_requested()