Bug#22384569: REDUCE THE SIZE OF THE JSON_WRAPPER CLASS
This patch makes the following changes that reduce the size of
Json_wrapper objects:

1) Remove the unused member variable m_id.

2) Move some member variables into a union, so that the objects don't
have room for both binary and DOM values at the same time. A
Json_wrapper either wraps a binary or a DOM, never both at the same
time, so it is enough to have space for one representation at a time.

3) Rename Json_wrapper::to_value() to to_binary(), and make it write
the binary representation directly into a supplied String instead of
returning a json_binary::Value. This makes the m_tmp member variable
unused so that it can be removed. Additionally, it saves one copying
of the binary representation when storing a value in a JSON column.