Bug #22070021	CHANGE VIRTUAL INDEX CALLBACK FUNCTION TO 		AVOID THE UNINTENDED TABLE OBJECT.
Problem:
======
The crash occurs due to we use the wrong TABLE object for computing
the virtual column. When a query uses multiple instances of the same table,
the current code will pick one of the open table objects, possibly not the
intended one.

Solution:
========
Change the interface for one of the call back functions
that InnoDB uses to get virtual column values.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
Reviewed-by: Marko Mäkelä <marko.makela@oracle.com>
RB: 11116