Failure of alter_table test after server merge
The alter_table test started failing after the recent server merge.
The failure was a result of Bug #20755615:
CREATING INDEX ON A RENAMED COLUMN WITH CASE CRASH .FRM FILE which
was fixed in commit e05558ba234cad2087de32709cd3a50d4e7f08cc.

The fix in the above commit involved making the check for change in
column name in InnoDB tables, if any, case sensitive. The flag being
set was being modified in cluster specific code. Included the
fix in cluster code as well to make sure the required change is
propagated and that the test case passes.