Bug#22366102: EMPTY STRINGS AS KEYS PROBABLY NOT HANDLED CORRECTLY BY JSON_EXTRACT()
If one of the path legs in a JSON path has an empty key name, such as
'$.""', a syntax error is raised.

Conversely, if JSON_SEARCH finds a match in a member with an empty key
name, it returns an invalid path, such as '$.' instead of the
correctly quoted path '$.""'.

The patch fixes this by moving the check for empty key names from
Json_path::parse_member_leg() to is_ecmascript_identifier(). This way,
we only check for empty key names when the path leg is unquoted, in
which case it has to follow the rules for ECMAScript identifiers, and
they cannot be empty. Empty key names will now be accepted if the path
leg is quoted. They are still rejected if they are unquoted.

Checking for empty strings in is_ecmascript_identifier() also fixes
the bug in JSON_SEARCH. Now it detects that the empty key name is not
a valid ECMAScript identifier, and adds the missing quotes around the
name.