Bug #22328395: REGRESSION: ?UNKNOWN COLUMN? FOR OUTER                COMPUTED VALUES USED INSIDE A SUBQUERY
ISSUE:
------
In the fix for Bug# 19823076, a subquery in the SELECT list
was no longer allowed to refer to an alias of a column in
the outer query.

This was justified, since it was not supported by the
standard sql syntax.

SOLUTION:
---------
Given that this is an extension used frequently, the
restriction introduced in Bug# 19823076 is now removed.