Bug#22321338 SELECTIVE INSTALL FOR DEVELOPMENT DOESN'T INSTALL ALL REQUIRED FILES
Trying to install just the shared libraries, clients, and development support
fails to install everything that is required.  After installing the
"SharedLibraries", "Client", and "Development" components it should be
possible to build a client application without errors, but this is NOT
currently possible.

Fix: add missing COMPONENT Development to binary_log_types.h