Followup patch for bug#22135584 (yet another)
The previous follow up patch for this bug introduced the
function find_epoch_to_handle(). Unfortunately that function
did not correctly handle that we could have a 's_pOp-event'
with an epoch number higher than the highest epoch received
on the 'i_pOp' upto now.

Thus we could end up handling an event from the s_pOp with a
higher epoch number than we would receive the next time we
polled for i_pOp event. This caused the schema vs binlog events
to be handled and logged logged in the incorrect relative order
(Which was the original problem described in this bug report)