Patch for fixing the failing MTR test ndb_rpl_conflict_epoch_extended_exceptions_table.test
The table mysql.ndb_replication has to be set up and
maintained in Mysql Cluster setup doing master-master replication.

In the final stages of the test
ndb_rpl_conflict_epoch_extended_exceptions_table.test it does
some 'Cleanup' where the test tables are dropped, and
the replication of the cleanup operations are checked.

However, also the ndb_replication table was dropped as
part of this cleanup - Which prevented the replication
itself to operate as intended.

Fix is to drop the ndb_replication table even later
when we tear down the slave.