BUG#21842399 reset slave all is not thread-safe
DESCRIPTION
===========
RESET SLAVE ALL could delete a channel, even master_pos_wait and
wait_until_sql_thread_after_gtid were still waiting for binlog to be applied.
It caused server crash when the functions tried to access the channel which
has been deleted.

FIX
===
A channel reference counter is added. It is increased if the channel
should not be deleted when the functions are running. RESET SLAVE ALL will wait
for no reference and then it can delete the channel.