Bug#22186926: CONVERT_CONSTANT_ITEM(THD*, ITEM_FIELD*, ITEM**): ASSERTION `!RESULT' FAILED.
Problem statement uses MAX/MIN optimization in opt_sum_query.
Index access is used to check if optmization is possible and
datetime column is not present in the index. So datetime field
is unread after optmization. In convert_constant_item()
incorrect value is obtained from the field since field is not
evaluated. Attempt to store this incorrect value to the field
later leads to assert failure.