Make testcase added as part of patch for bug#22204186 predictable.
The testcase issued a 'DUMP 900' command which caused all datanodes
to disconnect all API-clients. After a short while they would then reconnect.

The purpose was to create a condition where the binlog would temporary
bring all tables into a read only mode. The testcase would then
'wait-until' it was able to successfully update a row in a table as
an indication of the temp read only mode has ended.

This testing had the following problem:

1) The wait_until_not_readonly could possibly successfully complete
   before the Binlog thread had handled the datanode disconnect bringing
   the table into read only mode. (-> A false 'pass' of the test)

2) The disconnect could possibly not have happened yet when the test
   started the 'wait_until_not_readonly'. As the included test script
   was not prepared to see either the error codes 4010 or 4025 indicating
   an unexpected disconnect, the test failed.

This patch introduce a wait for a binlog rotate as an indicator
for the disconnect to have been seen by the API nodes.
It is then safe to wait for tables to become writable without any
more unexpected disconnects to happen.