Bug#22346120: GCOLS: CRASH IN ITEM_FUNC::WALK() WITH XOR A YEAR COLUMN
Re-evaluation of a generated column expression could access stale data
from the previous evaluation and lead to a server exit.

This is the same problem as bug#21810529, but the fix for that bug
only covered the TIME, DATE, DATETIME and TIMESTAMP data types. The
problem is still present for the YEAR data type.

The fix for bug#21810529 changed convert_constant_item() to skip the
conversion if the item had already been converted. One of the
conditions used for detecting that the item had been converted, was
that the item already was of the type that it should be converted to.

When a constant item is converted to TIME, DATE, DATETIME or
TIMESTAMP, the converted item is an Item_temporal_with_ref whose
field_type() function returns the type the item was converted to. If
it is converted to YEAR, on the other hand, the converted item is an
Item_int_with_ref whose field_type() function always returns
MYSQL_TYPE_LONGLONG. Since the type of the converted item is not
MYSQL_TYPE_YEAR, convert_constant_item() always thinks the item has
not yet been converted, and it tries to convert it again.

The second conversion could happen in a different memroot with a
shorter life than the generated column expression. If the generated
column expression outlives the converted item, re-evaluation of the
generated column expression will access a dead object and the server
will exit.

This patch moves the code that registers the type of the converted
item from Item_temporal_with_ref to its super-class Item_int_with_ref.
This change makes the converted item's field_type() function return
MYSQL_TYPE_YEAR instead of MYSQL_TYPE_LONGLONG in the case where the
constant item is converted to YEAR, which again makes the fix for
bug#21810529 work for the YEAR data type too.